Fixed a bug
The custom default directory would always be overwritten every time the user clicks the 'Apply' button in the settings menu. I fixed it so that it the directory path is only overwritten if `cfg.use_default_directory` is `False` or if the custom default directory path is not an empty string.